Removed addToWastePile method in FortyThievesWaste class.
According to the test requirements on Hertz's website:
Forty Thieves waste pile correctly returns if adding a specific card is
legal or illegal (e.g., always illegal since we use another approach for
this)

I think the other approach is the remove method from Stock pile, which
adds the removed stock pile card to the waste pile. I changed tests to
reflect this.